+++ src/utils/escapes.js
@@ -1,0 +1,11 @@
+// JSON string escape sequences - see json.org
+module.exports = {
+    '"': '"',
+    '\\': '\\',
+    '/': '/',
+    'b': '\b',
+    'f': '\f',
+    'n': '\n',
+    'r': '\r',
+    't': '\t'
+};