+++ CHANGELOG.md
@@ -229,7 +229,7 @@ No breaking changes have been made to the JSONata language itself.
 
 #### 1.2.1 Maintenance Release
 
-- Generate ES5 compatible version of jsonata.js (jsonata-es5.js & jsonata-es5.min.js) using Babel
+- Generate ES5 compatible version of flashteval.js (jsonata-es5.js & jsonata-es5.min.js) using Babel
     - use the -es5 version if you need to support older browsers or phantom.js
 - New functions $base64encode & base64decode
 - Fix formatting of inserts in error messages